Treads_Parkhaus update 4

+++ src/Threads_Parkhaus/ParkingGarage.java
@@ -2,7 +2,7 @@ package Threads_Parkhaus;
 
 public class ParkingGarage {
 
-    int parkplace;
+    long parkplace;
 
     public void enter(){
